server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
PyGithub==1.59.1
langchain==0.1.0
google-generativeai==0.3.1
langchain-google-genai==0.0.3
protobuf==4.23.4
huggingface_hub==0.16.4
pandas==2.0.3
numpy==1.25.2
python-docx==0.8.11
fpdf==1.7.2
markdown==3.4.4
gunicorn==21.2.0 